wip: implement selecting submatches
